+++ tests/lang.c
@@ -1,6 +1,6 @@
 #include <metalang99/assert.h>
 
-ML99_ASSERT_EMPTY_UNEVAL(ML99_eval(v()));
+ML99_ASSERT_EMPTY_UNEVAL(ML99_EVAL(v()));
 
 // A function with zero arguments {
 #define F_IMPL() v(123)
@@ -30,7 +30,7 @@ ML99_ASSERT_EQ(ML99_call(ML99_call(F, v(B, A, R)), v(6, 11)), v(6 + 11));
 #define F_IMPL()    v(1), v(2), v(3)
 #define BAR_IMPL(x) v()
 
-ML99_eval(ML99_call(BAR, ML99_call(F, v())))
+ML99_EVAL(ML99_call(BAR, ML99_call(F, v())))
 
 #undef F_IMPL
 #undef BAR_IMPL
@@ -50,7 +50,7 @@ ML99_ASSERT_EQ(ML99_call(F, v(OP)), v(123));
 // ML99_abort {
 ML99_ASSERT_EMPTY(ML99_abort(v()));
 ML99_ASSERT_EQ(ML99_abort(v(815057)), v(815057));
-ML99_ASSERT_UNEVAL(ML99_eval(v(~), ML99_abort(v(123)), v(~)) == 123);
+ML99_ASSERT_UNEVAL(ML99_EVAL(v(~), ML99_abort(v(123)), v(~)) == 123);
 
 // Ensure that `ML99_abort` also works correctly after some evaluations.
 #define F_IMPL() ML99_call(G, v(1, 2), ML99_call(H, v(123)))
@@ -58,7 +58,7 @@ ML99_ASSERT_UNEVAL(ML99_eval(v(~), ML99_abort(v(123)), v(~)) == 123);
     ML99_abort(v(ML99_ASSERT_UNEVAL(_1 == 1 && _2 == 2 && _123_plus_1 == 123 + 1)))
 #define H_IMPL(a) v(a + 1)
 
-ML99_eval(ML99_call(F, v()));
+ML99_EVAL(ML99_call(F, v()));
 
 #undef F_IMPL
 #undef G_IMPL
@@ -96,9 +96,9 @@ ML99_ASSERT_EQ(ML99_appl4(v(F), v(10), v(5), v(7), v(8)), v(10578));
 // The maximum arity {
 #define F_ARITY 255
 
-ML99_eval(ML99_consume(ML99_appl(v(F), v(~))))
-ML99_eval(ML99_consume(ML99_appl2(v(F), v(~), v(~))))
-ML99_eval(ML99_consume(ML99_appl3(v(F), v(~), v(~), v(~))))
+ML99_EVAL(ML99_consume(ML99_appl(v(F), v(~))))
+ML99_EVAL(ML99_consume(ML99_appl2(v(F), v(~), v(~))))
+ML99_EVAL(ML99_consume(ML99_appl3(v(F), v(~), v(~), v(~))))
 
 #undef F_ARITY
 // }