+++ tests/tuple.c
@@ -4,7 +4,7 @@
 #define CHECK(x, y) ML99_ASSERT_UNEVAL(x == 518 && y == 1910)
 
 // ML99_tuple {
-ML99_eval(v(CHECK), ML99_tuple(v(518, 1910)));
+ML99_EVAL(v(CHECK), ML99_tuple(v(518, 1910)));
 // }
 
 #undef CHECK
@@ -66,7 +66,7 @@ ML99_ASSERT_EQ(ML99_tupleGet(7)(v((19, 8, 7378, 10, 121, 1, 80, 23, 7))), v(23))
 #define CHECK_TAIL(...)            CHECK_TAIL_AUX(__VA_ARGS__)
 #define CHECK_TAIL_AUX(a, b, c, d) ML99_ASSERT_UNEVAL(a == 51 && b == 21 && c == 1 && d == 7378)
 
-CHECK_TAIL(ML99_eval(ML99_tupleTail(v((9191, 51, 21, 1, 7378)))));
+CHECK_TAIL(ML99_EVAL(ML99_tupleTail(v((9191, 51, 21, 1, 7378)))));
 
 #undef CHECK_TAIL
 #undef CHECK_TAIL_AUX
@@ -75,8 +75,8 @@ CHECK_TAIL(ML99_eval(ML99_tupleTail(v((9191, 51, 21, 1, 7378)))));
 // ML99_tupleAppend, ML99_tuplePrepend {
 #define CHECK(a, b, c) ML99_ASSERT_UNEVAL(a == 1 && b == 2 && c == 3)
 
-ML99_eval(v(CHECK), ML99_tupleAppend(ML99_tuple(v(1)), v(2, 3)));
-ML99_eval(v(CHECK), ML99_tuplePrepend(ML99_tuple(v(3)), v(1, 2)));
+ML99_EVAL(v(CHECK), ML99_tupleAppend(ML99_tuple(v(1)), v(2, 3)));
+ML99_EVAL(v(CHECK), ML99_tuplePrepend(ML99_tuple(v(3)), v(1, 2)));
 
 #undef CHECK
 // }